pyproject.toml
[tool.poetry]
name = "fastspeech2-lightning"
version = "0.1.0"
description = "Fastspeech2 training script based on lightning"
authors = ["Wataru Nakata <wataru9871@gmail.com>"]
readme = "README.md"
packages = [{include = "fastspeech2_lightning"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
hydra-core = "^1.3.2"
pyworld = "^0.3.2"
webdataset = "^0.2.39"
scipy = "^1.10.1"
pyrootutils = "^1.0.4"
torch = "^2.0.0"
torchaudio = "^2.0.1"
scikit-learn = "^1.2.2"
torchdata = "^0.6.0"
torchtext = "^0.15.1"
fastapi = "^0.80.0"
lightning = "^2.0.1.post0"
matplotlib = "^3.7.1"
wandb = "^0.14.2"
tensorboard = "^2.12.2"
torchvision = "^0.15.1"
soundfile = "^0.12.1"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
